requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
black==20.8b1
click==7.1.2
colorama==0.4.4
Django==3.1.5
djangorestframework==3.12.2
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==20.8
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytz==2020.5
regex==2020.11.13
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
